refactor(models): split checkpoint-loading policy out of BaseModel
BaseModel now owns only the runtime forward contract (padded/packed/VLM/CP/R3
input preparation, backbone forward, dense output restore). Everything about
turning a path into a distributed AutoModel module — registry probing, MoE
detection, kernel-backend selection, MTP disabling, fail-fast validation, and
post-load configuration — moves to molt/models/loading.py behind two entry
points: load_automodel() and configure_loaded_model().
