+++ tests/unit/test_model_capabilities.py
@@ -6,8 +6,9 @@ from types import SimpleNamespace
 import pytest
 from torch import nn
 
-import molt.models.base as base_module
-from molt.models.base import BaseModel, _automodel_supports_thd_packing
+import molt.models.loading as loading_module
+from molt.models.base import BaseModel
+from molt.models.loading import _automodel_supports_thd_packing
 
 
 class _DeclaredTHDModel(nn.Module):
@@ -101,9 +102,9 @@ def test_loaded_hf_fallback_configures_indexed_mask_packing(monkeypatch):
             load_kwargs.update(kwargs)
             return model
 
-    monkeypatch.setattr(base_module, "_detect_moe_arch", lambda _model: False)
-    monkeypatch.setattr(base_module, "_will_use_hf_model", lambda _path: True)
-    monkeypatch.setattr(base_module, "_mtp_off_kwargs", lambda _path: {})
+    monkeypatch.setattr(loading_module, "_detect_moe_arch", lambda _model: False)
+    monkeypatch.setattr(loading_module, "_will_use_hf_model", lambda _path: True)
+    monkeypatch.setattr(loading_module, "_mtp_off_kwargs", lambda _path: {})
     monkeypatch.setattr(utils, "convert_to_torch_dtype", lambda _dtype: None)
     monkeypatch.setattr(utils, "is_vlm_model", lambda _path: False)
     monkeypatch.setattr(nemo_automodel, "NeMoAutoModelForCausalLM", _FakeAutoModel)
@@ -127,16 +128,16 @@ def test_hf_indexed_mask_packing_rejects_cp_and_pp(monkeypatch, mesh_name):
         def __getitem__(self, _name):
             return SimpleNamespace(size=lambda: 2)
 
-    monkeypatch.setattr(base_module, "_detect_moe_arch", lambda _model: False)
-    monkeypatch.setattr(base_module, "_will_use_hf_model", lambda _path: True)
+    monkeypatch.setattr(loading_module, "_detect_moe_arch", lambda _model: False)
+    monkeypatch.setattr(loading_module, "_will_use_hf_model", lambda _path: True)
 
     with pytest.raises(NotImplementedError, match="cp_size=1 and pp_size=1"):
         BaseModel("dense-model", packing_samples=True, device_mesh=_Mesh())
 
 
 def test_hf_indexed_mask_packing_rejects_non_fa2(monkeypatch):
-    monkeypatch.setattr(base_module, "_detect_moe_arch", lambda _model: False)
-    monkeypatch.setattr(base_module, "_will_use_hf_model", lambda _path: True)
+    monkeypatch.setattr(loading_module, "_detect_moe_arch", lambda _model: False)
+    monkeypatch.setattr(loading_module, "_will_use_hf_model", lambda _path: True)
 
     with pytest.raises(ValueError, match="flash_attention_2"):
         BaseModel("dense-model", packing_samples=True, attn_implementation="flash_attention_3")
